Find new bounty on the list. Extract test support.
