+++ packages/apps/public/locales/en/test-support.json
@@ -1,0 +1,1 @@
+{}
