+++ packages/page-accounts/package.json
@@ -16,7 +16,6 @@
     "file-saver": "^2.0.5"
   },
   "devDependencies": {
-    "@testing-library/react": "^11.2.2",
-    "testcontainers": "^6.4.0"
+    "@polkadot/test-support": "0.72.2-5"
   }
 }
