+++ packages/test-support/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@polkadot/test-support",
+  "version": "0.72.2-5",
+  "main": "index.js",
+  "repository": "github:polkadot-js/apps",
+  "author": "Jaco Greeff <jacogr@gmail.com>",
+  "maintainers": [],
+  "contributors": [],
+  "license": "Apache-2.0",
+  "dependencies": {
+    "@testing-library/react": "^11.2.2",
+    "testcontainers": "^6.4.0"
+  }
+}
