+++ packages/test-support/src/keyring/MemoryStore.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2020 @polkadot/app-accounts authors & contributors
+// Copyright 2017-2020 @polkadot/test-support authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import type { KeyringJson, KeyringStore } from '@polkadot/ui-keyring/types';
