+++ packages/test-support/src/react/apiInTests.ts
@@ -1,0 +1,12 @@
+// Copyright 2017-2020 @polkadot/test-support authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import React, { PropsWithChildren } from 'react';
+
+import { useApi } from '@polkadot/react-hooks';
+
+export const WaitForApi = ({ children }: { children: React.ReactNode }): PropsWithChildren<any> | null => {
+  const api = useApi();
+
+  return api.isApiReady ? (children) : null;
+};
