+++ packages/test-support/src/react/index.ts
@@ -1,0 +1,4 @@
+// Copyright 2017-2020 @polkadot/test-support authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './apiInTests';
