+++ packages/test-support/src/substrate/index.ts
@@ -1,0 +1,4 @@
+// Copyright 2017-2020 @polkadot/test-support authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './constants';
